telegram_bot/requirements.txt
APScheduler==3.11.0
httpx==0.28.1
python-telegram-bot==21.10
PyYAML==6.0.2
